Extract Bib Page Functional Complexity into Bib Container
In order to simplify functionality between the Page and it's logic, extracing the functionality to it's own component helps clarify what is needed for the page and what is needed to functionality.
